+++ src/contexts/AuthUIContextProvider/provider.tsx
@@ -20,6 +20,7 @@ export const AuthUIContextProvider: React.FC<React.PropsWithChildren<AuthUIConte
             enableInviteRegistration: props.enableInviteRegistration,
             showContactSupport: props.showContactSupport,
             enableResetPassword: props.enableResetPassword,
+            enableCreatePassword: props.enableCreatePassword,
             showRememberMe: props.showRememberMe,
             showCybersecurityBadge: props.showCybersecurityBadge,
             allowDebugMode: props.allowDebugMode,
@@ -40,6 +41,7 @@ export const AuthUIContextProvider: React.FC<React.PropsWithChildren<AuthUIConte
             accountAlreadyExistsScreen: props.accountAlreadyExistsScreen,
             registrationConfig: props.registrationConfig,
             disablePagerAnimation: props.disablePagerAnimation,
+            showLoginForm: props.showLoginForm,
         };
 
         return propsForContext;
@@ -56,6 +58,7 @@ export const AuthUIContextProvider: React.FC<React.PropsWithChildren<AuthUIConte
         props.registrationSuccessScreen,
         props.enableInviteRegistration,
         props.enableResetPassword,
+        props.enableCreatePassword,
         props.htmlEula,
         props.loginActions,
         props.loginFooter,
@@ -70,6 +73,7 @@ export const AuthUIContextProvider: React.FC<React.PropsWithChildren<AuthUIConte
         props.showCybersecurityBadge,
         props.showSelfRegistration,
         props.registrationConfig,
+        props.showLoginForm,
     ]);
 
     return <AuthUIContext.Provider value={memoizedProps}>{props.children}</AuthUIContext.Provider>;